Authorization Added
Added Custom Policies;
Added ForbidException;
Added UserContextService;
Restricted acces for endpoints;
Refactored code.

+++ RestaurantsManagement/Services/RestaurantService.cs
@@ -1,4 +1,6 @@
+using System.Security.Claims;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 
 public interface IRestaurantService
@@ -15,16 +17,22 @@ public class RestaurantService : IRestaurantService
     private readonly RestaurantDbContext _dbContext;
     private readonly IMapper _mapper;
     private readonly ILogger<RestaurantService> _logger;
+    private readonly IAuthorizationService _authorizationService;
+    private readonly IUserContextService _userContextService;
 
     public RestaurantService(
         RestaurantDbContext dbContext,
         IMapper mapper,
-        ILogger<RestaurantService> logger
+        ILogger<RestaurantService> logger,
+        IAuthorizationService authorizationService,
+        IUserContextService userContextService
     )
     {
         _dbContext = dbContext;
         _mapper = mapper;
         _logger = logger;
+        _authorizationService = authorizationService;
+        _userContextService = userContextService;
     }
 
     public RestaurantDto GetById(int id)
@@ -56,6 +64,7 @@ public class RestaurantService : IRestaurantService
     public int Create(CreateRestaurantDto dto)
     {
         var restaurant = _mapper.Map<Restaurant>(dto);
+        restaurant.CreatedById = _userContextService.GetUserId;
 
         _dbContext.Restaurants.Add(restaurant);
         _dbContext.SaveChanges();
@@ -72,6 +81,18 @@ public class RestaurantService : IRestaurantService
         if (restaurant is null)
             throw new NotFoundException("Restaurant not found");
 
+        var authorizationResult = _authorizationService
+            .AuthorizeAsync(
+                _userContextService.User,
+                restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Delete)
+            )
+            .Result;
+        if (!authorizationResult.Succeeded)
+        {
+            throw new ForbidException();
+        }
+
         _dbContext.Restaurants.Remove(restaurant);
         _dbContext.SaveChanges();
     }
@@ -82,6 +103,18 @@ public class RestaurantService : IRestaurantService
 
         if (restaurant is null)
             throw new NotFoundException("Restaurant not found");
+
+        var authorizationResult = _authorizationService
+            .AuthorizeAsync(
+                _userContextService.User,
+                restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Update)
+            )
+            .Result;
+        if (!authorizationResult.Succeeded)
+        {
+            throw new ForbidException();
+        }
 
         restaurant.Name = dto.Name;
         restaurant.Description = dto.Description;
